Revert "Merge pull request #9 from onlyfanshop/tai"
This reverts commit 1373004b6807ba9e81971a159fc7ef9b094d513b, reversing
changes made to 7bc123ea713f196f57c73680ab1c8c0103434e62.

+++ src/main/java/com/example/onlyfanshop_be/repository/CartRepository.java
@@ -1,10 +1,11 @@
 package com.example.onlyfanshop_be.repository;
 
+
 import com.example.onlyfanshop_be.entity.Cart;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CartRepository extends JpaRepository<Cart, Long> {
+public interface CartRepository extends JpaRepository<Cart, Integer> {
 }
