+++ src/main/java/com/example/onlyfanshop_be/repository/ProductRepository.java
@@ -1,5 +1,6 @@
 package com.example.onlyfanshop_be.repository;
 
+
 import com.example.onlyfanshop_be.entity.Product;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -9,7 +10,7 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product, Long>, JpaSpecificationExecutor<Product> {
+public interface ProductRepository extends JpaRepository<Product, Integer>, JpaSpecificationExecutor<Product> {
 
 }
